requirements.txt
backcall==0.1.0
colorama==0.3.9
crispy-forms-foundation==0.7.1
decorator==4.1.2
Django==2.2.20
django-crispy-forms==1.8.1
django-debug-toolbar				
django-localflavor==3.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
ipdb==0.13.9
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.11.0
ldap3==2.5
localflavor==1.8
mysql_connector_python==8.0.28
mysqlclient==1.4.6			  
openpyxl==2.5.5
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.7
Pygments==2.7.4
python-stdnum==1.11
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
wfastcgi==3.0.0